+++ build.gradle
@@ -47,6 +47,9 @@ dependencies {
 
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
 
+	// pdf 관련
+	implementation 'org.apache.pdfbox:pdfbox:2.0.24'
+
 }
 test {
 	useJUnitPlatform()
