+++ src/main/java/com/erp/ezen25/controller/PartnerController.java
@@ -1,8 +1,8 @@
 package com.erp.ezen25.controller;
 
 import com.erp.ezen25.dto.*;
-import com.erp.ezen25.entity.Brand;
-import com.erp.ezen25.entity.Contract;
+import com.erp.ezen25.dto.contractDTO.*;
+import com.erp.ezen25.dto.productDTO.ProductBnameListResponseDTO;
 import com.erp.ezen25.service.BrandService;
 import com.erp.ezen25.service.ContractService;
 import lombok.RequiredArgsConstructor;
@@ -16,8 +16,6 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 @Controller
@@ -112,7 +110,7 @@ public class PartnerController {
     }
 
     @PostMapping("/contract/contractAdd")
-    public String contractAdd (ContractAddRequestDTO addRequest,@RequestParam("contractSelect") MultipartFile mf) throws IOException {
+    public String contractAdd (ContractAddRequestDTO addRequest, @RequestParam("contractSelect") MultipartFile mf) throws IOException {
         contractService.addContract(addRequest, mf);
         return "redirect:/ezen25/brand/contract/contractList";
     }
@@ -122,5 +120,23 @@ public class PartnerController {
     public ResponseEntity<Void> contractDelete (@RequestParam("contractId")Long contractId) {
         contractService.contractDelete(contractId);
         return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/contract/contractModify/{contractId}")
+    public String contractModifyForm(@PathVariable("contractId") Long contractId, Model model) {
+        List<ProductBnameListResponseDTO> bList = contractService.bnameList();
+        List<ContractPnameListResponseDTO> pList = contractService.prodList();
+        ContractDetailResponseDTO contract = contractService.contractDetail(contractId);
+
+        model.addAttribute("bList",bList);
+        model.addAttribute("pList",pList);
+        model.addAttribute("contract",contract);
+
+        return "/ezen25/brand/contract/contractModify";
+    }
+    @PostMapping("/contract/contractModify/{contractid}")
+    public String contractModify(ContractModifyRequest updateRequest, @RequestParam("contractSelect") MultipartFile mf) throws IOException {
+        contractService.contractUpdate(updateRequest, mf);
+        return "redirect:/ezen25/brand/contract/contractList";
     }
 }
