+++ src/main/java/com/erp/ezen25/dto/productDTO/ProductSCateListResponseDTO.java
@@ -1,0 +1,13 @@
+package com.erp.ezen25.dto.productDTO;
+
+import com.erp.ezen25.queryMapping.SCategoryListMapping;
+import lombok.Getter;
+
+@Getter
+public class ProductSCateListResponseDTO {
+    private final String s_category;
+
+    public ProductSCateListResponseDTO(SCategoryListMapping scMap) {
+        this.s_category = scMap.getS_category();
+    }
+}
