+++ src/main/resources/application.properties
@@ -15,7 +15,7 @@ server.tomcat.max-http-form-post-size=10MB
 server.servlet.context-path=/
 server.tomcat.uri-encoding=UTF-8
 
-spring.thymeleaf.prefix=classpath:/templates/
+# spring.thymeleaf.prefix=classpath:/templates/
 spring.thymeleaf.suffix=.html
 
 # Security ??
@@ -29,4 +29,12 @@ spring.servlet.multipart.enabled=true
 spring.servlet.multipart.location=C:/Users/Admin/IdeaProjects/erp_project/src/main/resources/static/image/
 spring.servlet.multipart.max-request-size=50MB
 spring.servlet.multipart.max-file-size=30MB
+
+# livereload ??
+spring.devtools.restart.enabled=true
+spring.devtools.livereload.enabled=true
+spring.freemarker.cache=false
+spring.thymeleaf.cache=false
+spring.thymeleaf.prefix=file:src/main/resources/templates/
+spring.resources.static-locations=file:src/main/resources/static/
