+++ src/main/resources/templates/ezen25/brand/contract/contractAdd.html
@@ -5,7 +5,7 @@
       layout:decorate="~{/layout/layout}">
 <th:block layout:fragment="content">
     <!-- 여기에 생성할 페이지 코드 작성 -->
-    <main class="formStyle">
+    <main class="formStyle contractAdd">
         <div class="container-fluid px-4">
             <h1 class="mt-4">계약등록</h1>
             <ol class="breadcrumb mb-4">
@@ -20,23 +20,28 @@
                             입력란
                         </div>
                     </div>
-                    <form method="post" class="card-body" id="productAddForm" enctype="multipart/form-data">
-                        <div class="inputLine">
+                    <form method="post" class="card-body" id="contractAddForm" enctype="multipart/form-data">
+                        <div class="inputLine relative">
                             <label for="brandName">계약 업체</label><input type="text" class="form-control" id="brandName" name="brandName" readonly>
                             <select class="form-control" name="bNameList" id="bNameList">
                                 <option>--선택--</option>
-                                <option value="0">-직접입력-</option>
                                 <option th:each=" bList: ${bList}" th:value="${bList.brandId}">[[${bList.brandName}]]</option>
                             </select>
                             <input type="hidden" id="brand" name="brand">
+                            <div class="card bg-primary text-white prodBtn optionAddBtn">
+                                <div class="card-body resetLink" onclick="window.open('/ezen25/brand/register')">업체추가</div>
+                            </div>
                         </div>
-                        <div class="inputLine">
+                        <div class="inputLine relative">
                             <label for="productName">제품 이름</label><input type="text" class="form-control" id="productName" name="productName" readonly>
                             <select class="form-control" name="pNameList" id="pNameList">
                                 <option>--선택--</option>
                                 <option th:each=" product: ${pList}" th:value="${product.productId}">[[${product.productName}]]</option>
                             </select>
                             <input type="hidden" id="productInfo" name="productInfo">
+                            <div class="card bg-primary text-white prodBtn optionAddBtn">
+                                <div class="card-body resetLink" onclick="window.open('/ezen25/product/productAdd' ,'height=500p,width=500px')">품목추가</div>
+                            </div>
                         </div>
                         <div class="inputLine">
                             <label for="contractDate">계약 날짜</label><input type="date" class="form-control" id="contractDate" name="contractDate">
@@ -45,6 +50,7 @@
                             <label for="contractSelect">계약서 파일</label>
                             <input type="file" class="form-control" id="contractSelect" name="contractSelect" accept=".pdf">
                         </div>
+                        <div class="mb-2" style="text-align: right;">* 계약서는 pdf 양식으로 제출 바랍니다.</div>
                         <div class="btnGroup">
                             <div class="card bg-primary text-white mb-2 prodBtn">
                                 <div class="card-body resetLink registerBtn">계약 등록</div>
@@ -55,14 +61,6 @@
                         </div>
                     </form>
                 </div>
-                <aside class="sideCard">
-                    <div class="card imageCard">
-                        <div class="card-header">품목 이미지</div>
-                        <div class="card-body">
-                            <img class="image" src="/image/NoImg.png" alt="이미지"/>
-                        </div>
-                    </div>
-                </aside>
             </div>
         </div>
     </main>
@@ -101,7 +99,7 @@
 
         $(".registerBtn").on("click", (e) => {
             console.log("ddd");
-            $("#productAddForm").submit();
+            $("#contractAddForm").submit();
         })
     </script>
 </th:block>
