+++ src/main/resources/templates/layout/fragments/sidebar.html
@@ -17,8 +17,8 @@
                     </a>
                     <div class="collapse" id="collapse1" aria-labelledby="headingOne" data-bs-parent="#sidenavAccordion">
                         <nav class="sb-sidenav-menu-nested nav">
-                            <a class="nav-link" href="layout-static.html">상품 목록</a>
-                            <a class="nav-link" href="layout-sidenav-light.html">상품 등록</a>
+                            <a class="nav-link" th:href="@{/ezen25/product/productList}">상품 목록</a>
+                            <a class="nav-link" th:href="@{/ezen25/product/productAdd}">상품 등록</a>
                         </nav>
                     </div>
                     <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#collapse2" aria-expanded="false" aria-controls="collapse2">
@@ -42,8 +42,8 @@
                     </a>
                     <div class="collapse" id="collapse3" aria-labelledby="headingOne" data-bs-parent="#sidenavAccordion">
                         <nav class="sb-sidenav-menu-nested nav">
-                            <a class="nav-link" href="#">계약서 목록</a>
-                            <a class="nav-link" href="#">계약등록</a>
+                            <a class="nav-link" th:href="@{/ezen25/brand/contract/contractList}">계약서 목록</a>
+                            <a class="nav-link" th:href="@{/ezen25/brand/contract/contractAdd}">계약등록</a>
                         </nav>
                     </div>
                     <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#collapse4" aria-expanded="false" aria-controls="collapse4">
